move vector type out of crud into typedef
